wip - each entity is its own component and being used in the entitystepcard component
