+++ services/ui-src/src/components/index.ts
@@ -24,8 +24,11 @@ export { Card } from "./cards/Card";
 export { EmailCard } from "./cards/EmailCard";
 export { TemplateCard } from "./cards/TemplateCard";
 export { EntityStepCard } from "./cards/EntityStepCard";
-export { EntityCardByStepType } from "./cards/EntityCardByStepType";
 export { ObjectiveProgressCardBottomSection } from "./cards/ObjectiveProgressCardBottomSection";
+export { ObjectiveProgressCard } from "./cards/ObjectiveProgressCard";
+export { FundingSourcesCard } from "./cards/FundingSourcesCard";
+export { EvaluationPlanCard } from "./cards/EvaluationPlanCard";
+
 // drawers
 export { Drawer } from "./drawers/Drawer";
 export { ReportDrawer } from "./drawers/ReportDrawer";
